move infinite bus validator to separate package
