+++ tests/test_device_validator.py
@@ -21,12 +21,11 @@ from uuid import uuid4
 import pytest
 
 from d3a_interface.constants_limits import ConstSettings
-from d3a_interface.device_validator import (
-    validate_finite_diesel_generator,
-    validate_infinite_bus)
+from d3a_interface.device_validator import validate_finite_diesel_generator
 from d3a_interface.exceptions import D3ADeviceException
 from d3a_interface.validators.cep_validator import validate_commercial_producer
 from d3a_interface.validators.home_meter_validator import HomeMeterValidator
+from d3a_interface.validators.infinite_bus_validator import validate_infinite_bus
 from d3a_interface.validators.load_validator import validate_load_device
 from d3a_interface.validators.market_maker_validator import validate_market_maker
 from d3a_interface.validators.pv_validator import validate_pv_device
